fix: branch highlighting was going to end of file for HTML reports
With the logic in place, seems like `startLine` will always == `endLine`
so the condition will never be `false`. Coupled with the fact that
the original `endCol` from the `meta` object seems wrong (always less
than `startCol`), the hightlighting just went to the end of the file.

Fix this by removing the redundant logic and setting the
`endCol` to the end of line using `text.originalLength()` when
the given `endCol` is less than `startCol`