fix: parenthesize superClass on non-idetifier case
Parenthesize superClass except it is an identifier. The superClass parenthesization is introduced in 1a13d1b493740054530ee8354e06d4e21d2ad1ac of #106 to support conditional expression. However, the process will turn the superClass into an parenthesizdExpression even when the superClass, in most cases, is an identifier only. The transformed AST will obstruct the extendsNative utility of babel-plugin-transform-classes since the superClass is not static anymore. This commit work arounds the compatibility by parenthesizing on non-identifier case only, thus the conditional expression is still well parenthesized with the added instruments while the identifier, without need of any instruments, is untouched.