feat: New options coverageGlobalScope and coverageGlobalScopeFunc.
This new option allows modification of the function which finds the
global variable scope.  The default global scope is `this`, resulting in
`(new Function('return this))()` to retrieve the global scope.  This can
now be changed, for example `window.top` could be used to always use the
top level IFRAME in a web browser.

coverageGlobalScopeFunc allows disabling of the evaluated function, just
directly references the specified variable.  This can be used the deal
with CSP issues but should not be used without coverageGlobalScope as
`this` could have a different meaning.

Fixes #199